+++ src/plugin.ts
@@ -8,6 +8,9 @@ import {
 } from "./actions/rs3-gim-group-rank";
 import { Rs3LevelTracker } from "./actions/rs3-level-tracker";
 import { Rs3WildyEventTimer } from "./actions/rs3-wildy-event-timer";
+import { Rs3Goldberg } from "./actions/rs3-goldberg";
+import { startUpdateScheduler } from "./update-scheduler";
+import { checkForUpdate } from "./update-check";
 
 streamDeck.logger.setLevel(LogLevel.TRACE);
 streamDeck.actions.registerAction(new Rs3GimGroupRankAboveLinked());
@@ -16,4 +19,8 @@ streamDeck.actions.registerAction(new Rs3GimGroupRank());
 streamDeck.actions.registerAction(new Rs3GimGroupRankOther());
 streamDeck.actions.registerAction(new Rs3LevelTracker());
 streamDeck.actions.registerAction(new Rs3WildyEventTimer());
+streamDeck.actions.registerAction(new Rs3Goldberg());
+startUpdateScheduler(checkForUpdate, (message, data) =>
+	streamDeck.logger.info(`${message} ${JSON.stringify(data ?? {})}`)
+);
 streamDeck.connect();